requirements.txt
alembic==1.0.10
aylien-apiclient==0.7.0
aylien-news-api==1.0.0
bokeh==1.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
lxml==4.3.3
Mako==1.0.10
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.3
packaging==19.0
pandas==0.24.2
pandas-datareader==0.7.0
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
statistics==1.0.3.5
tornado==6.0.2
urllib3==1.24.3
Werkzeug==0.15.2
wordninja==0.1.5
wrapt==1.11.1
